ARCHIVE/requirements.txt
# Backend dependencies
pandas
numpy
aiohttp
asyncio
psycopg2
python-dotenv

# Streamlit dashboard
streamlit
plotly
requests